donation count added to service

+++ FundraisingProject/src/main/java/com/codingdojo/FundraisingProject/Repositories/DonorRepo.java
@@ -24,5 +24,5 @@ public interface DonorRepo extends CrudRepository<Donor, Long>{
 	
 	//donation count functions
 	@Query(value = "SELECT COUNT(DISTINCT donations.id) FROM donors LEFT JOIN donations ON donations.donor_id = donors.id WHERE donors.id = :donorid", nativeQuery = true)
-	Double donordoncount(@Param("donorid") Long id);
+	Integer donordoncount(@Param("donorid") Long id);
 }
